+++ src/child.h
@@ -19,6 +19,7 @@ extern void child_sendw(const wchar *, uint len);
 extern void child_resize(struct winsize * winp);
 extern bool child_is_alive(void);
 extern bool child_is_parent(void);
+extern char * procres(int pid, char * res);
 extern wchar * grandchild_process_list(void);
 extern char * child_tty(void);
 extern char * foreground_prog(void);  // to be free()d